requirements.txt
neurolib
matplotlib
seaborn
pytest
numpy
scipy
xarray
PyYAML
numba
tqdm